pyproject.toml
[tool.poetry]
name = "archon-ai"
version = "0.1.2"
description = ""
authors = ["Shlok Natarajan <shlok.natarajan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
pydantic = "^2.9.1"
openai = "^1.45.0"
anthropic = "^0.34.2"
groq = "^0.11.0"
google-generativeai = "^0.8.1"
litellm = "^1.45.0"
tiktoken = "^0.7.0"
requests = "^2.32.3"
tqdm = "^4.66.5"
datasets = "^3.0.1"
torch = "^2.4.1"
numpy = "^1.26.3"
pyyaml = "^6.0.2"
shortuuid = "^1.0.13"
pydra-config = "^0.0.2"
bayesian-optimization = "^1.5.1"
tabulate = "^0.9.0"
together = "^1.3.0"
transformers = "^4.45.1"
psutil = "^6.0.0"
accelerate = "^0.34.2"
plotly-express = "^0.4.1"
docker = "^7.1.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
fire = "^0.7.0"
nltk = "^3.9.1"
prettytable = "^3.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
